Revert "fix(renderer): repair canvas backing synchronously on reveal fit"
This reverts commit 60bfa6bed04586b0b1513b295c48b024b8b869ac.
