+++ src/renderer/src/components/terminal-pane/terminal-visibility-resume.ts
@@ -16,7 +16,7 @@ import {
 import { focusActivePane } from './pane-helpers'
 import { scheduleTabRevealWebglAtlasRecovery } from './terminal-webgl-atlas-recovery'
 import { flushDeferredPaneMetricOptionsIfMeasurable } from '@/lib/pane-manager/pane-fit'
-import { repairPaneWebglCanvasBackingMismatch } from '@/lib/pane-manager/terminal-canvas-backing-repair'
+import { repairPaneWebglCanvasDprMismatch } from '@/lib/pane-manager/terminal-canvas-dpr-repair'
 
 const VISIBLE_RESUME_FLUSH_CHARS = 256 * 1024
 const WINDOW_WAKE_FLUSH_CHARS = 64 * 1024
@@ -80,9 +80,8 @@ export function resumeTerminalVisibility({
           flushedDeferredMetrics = true
         }
         // Why here: the light path neither recreates WebGL nor fits, so a dpr
-        // or layout change that landed while this tab was hidden has no other
-        // repair point.
-        repairPaneWebglCanvasBackingMismatch(pane)
+        // change that landed while this tab was hidden has no other repair point.
+        repairPaneWebglCanvasDprMismatch(pane)
       }
       // Why: intra-worktree tab switches only toggle the overlay. Keeping
       // synchronous drain and atlas rebuilds off this path avoids racing the
